Completely rework frame rendering
Previously when ever some content was changed by the Qt Application, we were
updating the desktop image, and notify the peers about these changes. After this
refactoring, the content is updated with a framerate, and so every fps a timer is
triggered and looks at the changes since the last frame. In the meantime changes
are just accumulated, and we update things (and notify RDP peers) only at a certain
rate.