Various changes and improvements on config and pointers
This patch adds various neat options to control which channels are enabled
by adding noegfx and noclipboard parameters to the QPA.

A new "secrets" options allows to specify the secrets files to dump the SSL
secrets (very useful to see the traffic in clear when the client is mstsc).

It also reworks the handling of the dynamic channel so that we poll its state
at the right moment.

And finally the pointers are no longer sent using largePointer commands but
pointerNew commands instead. The pointer cache may also be handled more reliably.