Fixed egfx states and segfault when droppping a window
Mstsc was unhappy about the way we were treating the EGFX channel, so it was
closing the connection because we were not waiting for capabilities message
before sending some content. The patch introduce a new flag to accumulate changes
until the EGFX channel is correctly initialized.

Also fixed a segfault when the window widget had been the entered widget, but
the window was destroyed.