Add an option for generating layout independent Qt::Key in events
When confronted with an unknown platform like qfreerdp, qtwebengine
falls back to assuming that all inputs are generated by a US qwerty
keyboard.

In order to be able to use web apps that rely on DOM (key) .code
(supposed to represent the physical location of the keys on the
keyboard, not affected by any layout setting), we transform the Qt::Key
sent along each QKeyEvent as if sent by a US keyboard.

Given that qtwebengine seems to rely on the `text` attribute to generate
DOM .key event attributes, this workaround should not affect text entry
with different layouts.
https://github.com/qt/qtwebengine/blob/603a1809481eb4d4ca972f0f64915d29fb99f53b/src/core/web_event_factory.cpp#L1641