Refactor peer keyboard handling to its own class
QFreeRdpPeer was starting to get a bit too crowded, especially with all
the conversion tables. Keyboard handling fits well as its own unit, and
will make writing unit tests for it later easier.

This was also the occasion to review the code and upgrade some call to
libxkbcommon to a more modern API.