Refactoring and fixes
Do an incremental refactoring to prepare the project for GL introduction, multiple monitors
and dynamic resize.

Fix he handling of backingStore objects: they are associated with QWindow so when a new
platform window is created scan for a previous association. That fixes some bugs with
menus when sometime when the menu is shown, hidden then shown again the menu is not
visible because of a backing store re-use.