First window without decorations is not configurable
The first created window has a special treatment: we never decorate it (so no way to move or
resize it), and when you click on it, it is not raised on top.
With this new `norootwindow` parameter you can now disable this behaviour and all windows
including the first one are all treated the same.